Add amplify resolvers and admin config for sort in search
